+++ protocols/kad/src/behaviour.rs
@@ -589,7 +589,13 @@ where
 
         let done = records.len() >= quorum.get();
         let target = kbucket::Key::new(key.clone());
-        let info = QueryInfo::GetRecord { key: key.clone(), records, quorum, cache_at: None };
+        let info = QueryInfo::GetRecord {
+            key: key.clone(),
+            records,
+            quorum,
+            cache_at: None,
+            no_record: Vec::new(),
+        };
         let peers = self.kbuckets.closest_keys(&target);
         let inner = QueryInner::new(info);
         let id = self.queries.add_iter_closest(target.clone(), peers, inner); // (*)
@@ -602,7 +608,8 @@ where
         id
     }
 
-    /// Stores a record in the DHT.
+    /// Stores a record in the DHT, locally as well as at the nodes
+    /// closest to the key as per the xor distance metric.
     ///
     /// Returns `Ok` if a record has been stored locally, providing the
     /// `QueryId` of the initial query that replicates the record in the DHT.
@@ -636,6 +643,54 @@ where
         };
         let inner = QueryInner::new(info);
         Ok(self.queries.add_iter_closest(target.clone(), peers, inner))
+    }
+
+    /// Stores a record at specific peers, without storing it locally.
+    ///
+    /// The given [`Quorum`] is understood in the context of the total
+    /// number of distinct peers given.
+    ///
+    /// If the record's expiration is `None`, the configured record TTL is used.
+    ///
+    /// > **Note**: This is not a regular Kademlia DHT operation. It may be
+    /// > used to selectively update or store a record to specific peers
+    /// > for the purpose of e.g. making sure these peers have the latest
+    /// > "version" of a record or to "cache" a record at further peers
+    /// > to increase the lookup success rate on the DHT for other peers.
+    /// >
+    /// > In particular, if lookups are performed with a quorum > 1 multiple
+    /// > possibly different records may be returned and the standard Kademlia
+    /// > procedure of "caching" (i.e. storing) a found record at the closest
+    /// > node to the key that _did not_ return it cannot be employed
+    /// > transparently. In that case, client code can explicitly choose
+    /// > which record to store at which peers for analogous write-back
+    /// > caching or for other reasons.
+    pub fn put_record_to<I>(&mut self, mut record: Record, peers: I, quorum: Quorum) -> QueryId
+    where
+        I: ExactSizeIterator<Item = PeerId>
+    {
+        let quorum = if peers.len() > 0 {
+            quorum.eval(NonZeroUsize::new(peers.len()).expect("> 0"))
+        } else {
+            // If no peers are given, we just let the query fail immediately
+            // due to the fact that the quorum must be at least one, instead of
+            // introducing a new kind of error.
+            NonZeroUsize::new(1).expect("1 > 0")
+        };
+        record.expires = record.expires.or_else(||
+            self.record_ttl.map(|ttl| Instant::now() + ttl));
+        let context = PutRecordContext::Custom;
+        let info = QueryInfo::PutRecord {
+            context,
+            record,
+            quorum,
+            phase: PutRecordPhase::PutRecord {
+                success: Vec::new(),
+                get_closest_peers_stats: QueryStats::empty()
+            }
+        };
+        let inner = QueryInner::new(info);
+        self.queries.add_fixed(peers, inner)
     }
 
     /// Removes the record with the given key from _local_ storage,
@@ -1083,7 +1138,7 @@ where
                 }
             }
 
-            QueryInfo::GetRecord { key, records, quorum, cache_at } => {
+            QueryInfo::GetRecord { key, records, quorum, cache_at, no_record } => {
                 let results = if records.len() >= quorum.get() { // [not empty]
                     if let Some(cache_key) = cache_at {
                         // Cache the record at the closest node to the key that
@@ -1103,7 +1158,7 @@ where
                         let inner = QueryInner::new(info);
                         self.queries.add_fixed(iter::once(cache_key.into_preimage()), inner);
                     }
-                    Ok(GetRecordOk { records })
+                    Ok(GetRecordOk { records, no_record })
                 } else if records.is_empty() {
                     Err(GetRecordError::NotFound {
                         key,
@@ -1153,7 +1208,7 @@ where
                     }
                 };
                 match context {
-                    PutRecordContext::Publish =>
+                    PutRecordContext::Publish | PutRecordContext::Custom =>
                         Some(KademliaEvent::QueryResult {
                             id: query_id,
                             stats: get_closest_peers_stats.merge(result.stats),
@@ -1252,7 +1307,7 @@ where
                     }
                 });
                 match context {
-                    PutRecordContext::Publish =>
+                    PutRecordContext::Publish | PutRecordContext::Custom =>
                         Some(KademliaEvent::QueryResult {
                             id: query_id,
                             stats: result.stats,
@@ -1722,7 +1777,7 @@ where
             } => {
                 if let Some(query) = self.queries.get_mut(&user_data) {
                     if let QueryInfo::GetRecord {
-                        key, records, quorum, cache_at
+                        key, records, quorum, cache_at, no_record
                     } = &mut query.inner.info {
                         if let Some(record) = record {
                             records.push(PeerRecord{ peer: Some(source), record });
@@ -1744,19 +1799,22 @@ where
                                     );
                                 }
                             }
-                        } else if quorum.get() == 1 {
-                            // It is a "standard" Kademlia query, for which the
-                            // closest node to the key that did *not* return the
-                            // value is tracked in order to cache the record on
-                            // that node if the query turns out to be successful.
-                            let source_key = kbucket::Key::from(source);
-                            if let Some(cache_key) = cache_at {
-                                let key = kbucket::Key::new(key.clone());
-                                if source_key.distance(&key) < cache_key.distance(&key) {
+                        } else {
+                            no_record.push(source);
+                            if quorum.get() == 1 {
+                                // It is a "standard" Kademlia query, for which the
+                                // closest node to the key that did *not* return the
+                                // value is tracked in order to cache the record on
+                                // that node if the query turns out to be successful.
+                                let source_key = kbucket::Key::from(source);
+                                if let Some(cache_key) = cache_at {
+                                    let key = kbucket::Key::new(key.clone());
+                                    if source_key.distance(&key) < cache_key.distance(&key) {
+                                        *cache_at = Some(source_key)
+                                    }
+                                } else {
                                     *cache_at = Some(source_key)
                                 }
-                            } else {
-                                *cache_at = Some(source_key)
                             }
                         }
                     }
@@ -2063,7 +2121,11 @@ pub type GetRecordResult = Result<GetRecordOk, GetRecordError>;
 /// The successful result of [`Kademlia::get_record`].
 #[derive(Debug, Clone)]
 pub struct GetRecordOk {
-    pub records: Vec<PeerRecord>
+    /// The records found, including the peer that returned them.
+    pub records: Vec<PeerRecord>,
+    /// Peers that were queried but did not return a record
+    /// for the queried key.
+    pub no_record: Vec<PeerId>,
 }
 
 /// The error result of [`Kademlia::get_record`].
@@ -2319,17 +2381,32 @@ impl QueryInner {
 /// The context of a [`QueryInfo::AddProvider`] query.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub enum AddProviderContext {
+    /// The context is a [`Kademlia::start_providing`] operation.
     Publish,
+    /// The context is periodic republishing of provider announcements
+    /// initiated earlier via [`Kademlia::start_providing`].
     Republish,
 }
 
 /// The context of a [`QueryInfo::PutRecord`] query.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub enum PutRecordContext {
+    /// The context is a [`Kademlia::put_record`] operation.
     Publish,
+    /// The context is periodic republishing of records stored
+    /// earlier via [`Kademlia::put_record`].
     Republish,
+    /// The context is periodic replication (i.e. without extending
+    /// the record TTL) of stored records received earlier from another peer.
     Replicate,
+    /// The context is an automatic write-back caching operation of a
+    /// record found via [`Kademlia::get_record`] at the closest node
+    /// to the key queried that did not return a record. This only
+    /// occurs after a lookup quorum of 1 as per standard Kademlia.
     Cache,
+    /// The context is a custom store operation targeting specific
+    /// peers initiated by [`Kademlia::put_record_to`].
+    Custom,
 }
 
 /// Information about a running query.
@@ -2389,6 +2466,8 @@ pub enum QueryInfo {
         records: Vec<PeerRecord>,
         /// The number of records to look for.
         quorum: NonZeroUsize,
+        /// The peers what were queried but did not return a record for the `key`.
+        no_record: Vec<PeerId>,
         /// The closest peer to `key` that did not return a record.
         ///
         /// When a record is found in a standard Kademlia query (quorum == 1),